Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int value;
+    float value;
     printf("input :");
     scanf("%f",&value);
-    printf("You entered ; %d" , value)
+    printf("You entered ; %f" , value)
     return 0;
 }
